\name{SoyNAM-package}
\alias{SoyNAM-package}
\docType{package}
\title{
Soybean Nested Association Mapping Dataset
}
\description{
    Genomic and multi-environmental soybean data. Soybean Nested
    Association Mapping (SoyNAM) project dataset funded by the United Soybean Board
    (USB). BLUP function formats data for genome-wide prediction and association analysis.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.6.2 \cr
Date: \tab 2022-01-03\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, William Beavis, James Specht, Brian Diers, Rouf Mian, Reka Howard,
George Graef, Randall Nelson, William Schapaugh, Dechun Wang, Grover Shannon, Leah McHale,
Perry Cregan, Qijian Song, Miguel Lopez, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}
\seealso{
Functions: BLUP, data(soybase), data(soynam) and data(soyin).
}
