\name{SoyNAM-package}
\alias{SoyNAM-package}
\docType{package}
\title{
Soybean Nested Association Mapping Dataset
}
\description{
Genomic and multi-environmental soybean data. Soybean Nested Association Mapping (SoyNAM) project dataset funded by the United Soybean Board (USB), pre-formatted for general analysis and genome-wide association analysis using the NAM package.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.1 \cr
Date: \tab 2015-10-15\cr
License: \tab GPL-3 \cr
}
}
\author{
Alencar Xavier, William Beavis, James Specht, Brian Diers, William Muir, Katy Rainey.

Maintainer: Alencar Xavier <xaviera@purdue.edu>
}
\seealso{
Functions: BLUP1, BLUP2, data(soybase) and data(soynam)
}